refactor(keyvault): split MyRolesSection into per-tab components
Move presence/status and profile into components/MyRolesStatusTab and
components/MyRolesProfileTab, and the shared tables into
myRolesConstants. MyRolesSection keeps only the tab switch and its
public exports (MY_ROLES_TAB, MyRolesTab, default). Behaviour unchanged.
